requirements.txt
anthropic==0.40.0
psycopg2-binary==2.9.9
pgvector==0.3.2
pdfplumber==0.11.0
pymupdf4llm
python-docx
openpyxl
markdownify
python-dotenv==1.0.1
rank-bm25==0.2.2
tiktoken==0.7.0
pydantic==2.6.0
# Sprint 2
fastapi==0.111.0
uvicorn==0.29.0
streamlit==1.35.0
httpx==0.27.0
beautifulsoup4>=4.12.0
python-multipart==0.0.9
pytest-asyncio==0.23.6
voyageai
# Admin Module
bcrypt==5.0.0
PyJWT==2.12.1
# Security
slowapi==0.1.9
limits==3.6.0
python-magic==0.4.27
